dynlevy scaling curve done, current levy efficiency shown in Laws
One should be able to view their current dynlevy law and its
associated levy-raising efficiency percentage (before the effects
of levy laws) along with an explanatory custom tooltip and law
modifiers in effect in the Laws view now.  This was accomplished
by making each law's potential be has_law = <this-law>, setting
the allow to false, and adding appropriate i18n.

Hopefully, this works, as it's a really neat way to be able to
see the current strategic effects of otherwise hidden laws. It
also makes integrity-checking as simple as switching to a ruler
and checking their Laws view.