Added user_dir to EMF.mod
Developers testing EMF w/o the aid of the installer should replace
the path to the mod to the relative path to the EMF git repository
on their local machine. That way, you can always test your changes
to the working tree, switch between branches, etc. and never have
to fiddle with copying files around or the like.