Added Matriarchal Court Spawning Mimicking Vanilla
For characters with a matriarchal religion, the `emf_create_courtiers` effect will now spawn female councillors, female commanders, and male bachelors. This mimics the way certain vanilla event outcomes work for matriarchal religions, so is deemed to be included with the compatch even if it seems like a tangential enhancement.