Fix *Varangian and the Princess* event chain (see msg)
As I noted previously in a GitHub commit comment, the scoping from
`d_varangian_guard` (title) straight to `any_realm_character` does
not work (I caught a lot of errors in the log while playing, which
confirmed it). So I simply scoped to `owner` in between the two.

I also removed the `prefers_men_trigger` checks, because the event
already guarantees that both the varangian and the "princess" are
not homosexuals and of the opposite sex.

While I was at it, I switched from using regular state opinion
(`opinion`) to `personal_opinion`, which is much more commonly
used for attraction purposes since WoL was released.

Also, the event was `only_playable = yes` and yet it can only fire
for an emperor, so quick MTTH event optimization is to use
`only_independent = yes`.

Further, I made it possible for the event to fire when the emperor
`is_incapable` -- in this case, however, they will not have the option
to execute the Varangian.