Emperor Advisor Prestige, Great Councillors
Re-added Great Councillors to the mod, this time implemented differently. Each patrician that isn't on the council unlocks a Great Councillor slot. Slots are usually automatically occupied, but they can be manually vacated by the doge in case they want to hire someone on as a proper councillor. There will be some cases where they will need to be manually given out, especially if a patrician gets promoted to a full councillor position that causes the number of Great Councillor positions to contract, firing the person who occupied the last slot.