Freckles Stuff
Also changed the way the freckles startup event works so that it's more legible (don't know if it performs better though). Instead of being a single event that applies freckles to all characters in one call, there is now a separate call for each character that then applies freckles to that character.