Removed Unnecessary `disallow_new_battle_events_effect` Calls
`disallow_new_battle_events_effect` and `allow_new_battle_events_effect` are meant to be used at the beginning and end of battle events called by the main tombola that fires on every combat pulse. Their inclusion in the main tomobola itself is therefore unnecessary and could potentially cause cases where the underlying flag is set and remains so for eternity.