Warlike Manpower Mult Fix, Removed Doctrine Feature Updater for Older Saves
Made the call to remove the event that updates the religious features of already-reformed, pagan religions in existing savegames. It is incredibly difficult to set the event up to make sure the proper features are set (mainly unit and character modifiers) and to only override doctrines in older saves. The event would also cause problems with submods that add, remove, or change religious features. The only thing lost is the fact that already-reformed religions in older savegames might end up with a mix of old and new features, but that's a relatively minor issue IMHO and shouldn't be game-breaking.