Swaying Changes, Partial Reversion of 4-Day Stalling Fix
Decided to revert back to `random_character` from `random_playable_ruler` because of the need to be able to sway key unlanded characters (e.g. councilors), but made sure to start its `limit` block with a set of conditions that should quickly filter out unnecessary unlanded characters before the more expensive conditions are checked. Playtesting revealed that there is a very minor stall every 4 days, but it's nowhere near as noticeable as before.