Reshuffled Combat Tactics Bonus Generation
The .csv files for bonuses for Cultural and special Call to Glory tactics now only store the extra bonuses on top of those of the original tactics they replace instead of all the bonuses. This should make it much easier to reflect updates in the base tactic changes to corresponding Cultural and/or special Call to Glory tactics.